Version 1 de transmicion

+++ src/app/component/tips/tips.component.ts
@@ -5,7 +5,6 @@ import { DomSanitizer } from "@angular/platform-browser";
 import { Router } from '@angular/router';
 import { Tips } from '../../interfaces/portal';
 import { PostService } from '../../services/post.service';
-import { SesionService } from '../../services/sesion.service';
 
 
 
@@ -17,39 +16,90 @@ import { SesionService } from '../../services/sesion.service';
 
 export class TipsComponent implements OnInit {
 
+  event: boolean = false;
 
-  
-  constructor(private router: Router, private sanitizer: DomSanitizer, private sesionService: SesionService, private postService: PostService, private dialogEvents: MatSnackBar) {
+  falta: string = "";
+
+  finish: boolean = false;
+
+  constructor(private router: Router, private sanitizer: DomSanitizer, private postService: PostService, private dialogEvents: MatSnackBar) {
 
   }
-
-
 
   @Input() tips: Tips = { url: "", content: "", id: "" };
 
   @Input() isDelete: boolean = false;
+
+  @Input() preEdit: boolean = false;
+
+  @Input() isOnline: boolean = false;
 
   configDialog: MatSnackBarConfig = {
     duration: 10000, verticalPosition: 'top'
   }
 
   ngOnInit(): void {
+    const hoy = new Date();
+
+    const tipsDate = this.tips.date ? new Date(this.tips.date) : hoy;
+    this.event = tipsDate > hoy;
+
+    if (this.event) {
+      this.falta = this.cuantoFalta(hoy, tipsDate);
+    }
+
   }
 
-  killme() {
-    this.gameOver(this.tips.id)
+  cuantoFalta(startDate: Date, endDate: Date): string {
+    const diffInMs = endDate.getTime() - startDate.getTime();
+    const diffInDays = Math.ceil(diffInMs / (1000 * 60 * 60 * 24));
+
+    if (diffInDays < 1) {
+      return 'Hoy';
+    } else if (diffInDays === 1) {
+      return 'Dentro de un día';
+    } else if (diffInDays < 7) {
+      return `Dentro de ${diffInDays} días`;
+    } else if (diffInDays < 30) {
+      const weeks = Math.ceil(diffInDays / 7);
+      return `Dentro de ${weeks} semana${weeks > 1 ? 's' : ''}`;
+    } else {
+      const months = Math.ceil(diffInDays / 30);
+      return `Dentro de ${months} mes${months > 1 ? 'es' : ''}`;
+    }
+  }
+
+  killme(): boolean {
+    if (this.finish && this.isDelete) {
+      this.gameOver(this.tips.id)
+      return true;
+    } else {
+      return false;
+    }
+  }
+
+  edit() {
+    if (this.isDelete) {
+      this.gameOver(this.tips.id)
+    }else{
+      this.dialogEvents.open("quieren hackear el sitioooo!!", "cerrar", this.configDialog);
+    }
   }
 
   gameOver(id: string): void {
-    this.postService.deletePost(id).subscribe(
-      () => {
-        this.dialogEvents.open("Noticia eliminada!", "cerrar", this.configDialog);
-        this.router.navigate(['/news']);
-      },
-      (error) => {
-        this.dialogEvents.open("No fue posible eliminar", "cerrar", this.configDialog);
-      }
-    );
+    if (this.isDelete) {
+      this.postService.deletePost(id).subscribe(
+        () => {
+          this.dialogEvents.open("Noticia eliminada!", "cerrar", this.configDialog);
+          this.router.navigate(['/dashboard']);
+        },
+        (error) => {
+          this.dialogEvents.open("No fue posible eliminar", "cerrar", this.configDialog);
+        }
+      );
+    }else{
+      this.dialogEvents.open("quieren hackear el sitioooo!!", "cerrar", this.configDialog);
+    }
   }
 
   shared() {
@@ -58,9 +108,9 @@ export class TipsComponent implements OnInit {
   }
 
 
-  msgText() : string{
-    var url : string = "https://" + document.location.hostname + "/s?id=" + this.tips.id
-    return this.tips.title + ", " + url; 
+  msgText(): string {
+    var url: string = "https://" + document.location.hostname + "/s?id=" + this.tips.id
+    return this.tips.title + ", " + url;
   }
